Split mobile and tv configs
There are a lot of apps installed by default that are unusable with
a controller. This splits the mobile related settings out of common
into a common_mobile config and adds common_tv.

Note: The EXCLUDE_AUDIOFX flag was removed from the tv configs
because the new config path no longer builds it.

Change-Id: I67db32750fca91725d703243f2bcee8bcd9b82cd